Chatbot Tư vấn Hướng nghiệp/requirements.txt
flask==2.3.2
faiss-cpu==1.7.4
numpy==1.26.0
google-generativeai==0.3.1
sentence-transformers==2.2.2
PyMuPDF==1.23.7
python-docx==1.0.1
nltk==3.8.1